+++ app/composables/useStaffAccountManagement.ts
@@ -1,0 +1,6 @@
+import { createFirebaseStaffAccountGateway } from '../gateways/staffAccountGateway.ts'
+
+export const useStaffAccountManagement = () => {
+  const { $firebase } = useNuxtApp()
+  return createFirebaseStaffAccountGateway($firebase.functions, $firebase.auth)
+}